Refactor server.py into kernel, session, and server modules
Split the monolithic server.py into focused modules:
- kernel.py: Jupyter kernel process management
- session.py: Session class and registry
- server.py: FastAPI routes and app setup
